+++ tests/utilities.test.ts
@@ -240,7 +240,7 @@ const testContext = {
   rdf: "http://www.w3.org/1999/02/22-rdf-syntax-ns#",
   rdfs: "http://www.w3.org/2000/01/rdf-schema#",
   xsd: "http://www.w3.org/2001/XMLSchema#",
-  dc: "http://purl.org/dc/terms/",
+  dcterms: "http://purl.org/dc/terms/",
   schema: "http://schema.org/",
   ex: "https://example-vocab.ex/",
   soa: "http://schema-org-adapter.at/vocabTerms/",
@@ -296,8 +296,8 @@ const testContext = {
     "@id": "schema:inverseOf",
     "@type": "@id",
   },
-  "dc:source": {
-    "@id": "dc:source",
+  "dcterms:source": {
+    "@id": "dcterms:source",
     "@type": "@id",
   },
   "schema:source": {
@@ -310,7 +310,7 @@ const testContext2 = {
   rdf: "http://www.w3.org/1999/02/22-rdf-syntax-ns#",
   rdfs: "http://www.w3.org/2000/01/rdf-schema#",
   xsd: "http://www.w3.org/2001/XMLSchema#",
-  dc: "http://purl.org/dc/terms/2",
+  dcterms: "http://purl.org/dc/terms/2",
   schema2: "http://schema.org/",
   ex: "https://example-vocab.ex/",
   soa: "http://schema-org-adapter.at/vocabTerms/",
@@ -366,8 +366,8 @@ const testContext2 = {
     "@id": "schema:inverseOf",
     "@type": "@id",
   },
-  "dc:source": {
-    "@id": "dc:source",
+  "dcterms:source": {
+    "@id": "dcterms:source",
     "@type": "@id",
   },
   "schema:source": {